Patch test and move simulator to main package
